+++ R/format.r
@@ -6,18 +6,17 @@
 #' `format_pdfmark()` returns a pdfmark datetime string with as much **known** information possible.
 #' `format_strftime()` allows [base::strftime()] style formatting.
 #' `format_nanotime()` allows CCTZ style formatting.
+#' `format_edtf()` returns an Extended Date Time Format (EDTF) string.
 #' @name format
 #' @param x A [datetimeoffset()] object.
 #' @param format For `format_strftime()` see [base::strftime()].
 #'               For `format_nanotime()` see <https://github.com/google/cctz/blob/6e09ceb/include/time_zone.h#L197>.
 #' @param tz A character string specifying the time zone to be used for the conversion.
 #'           Can be a length greater than one.
-#' @param usetz Include the time zone in the formatting (of outputs including
-#'        time; date outputs never include time zone information).
-#' @param precision The amount of precision to represent with substrings of
-#'        "ymdhmsn", as "y"ear, "m"onth, "d"ay, "h"our, "m"inute,
-#'        "s"econd, and "n"anosecond. (e.g. "ymdhm" would show precision through minutes.
-#'        When ‘NULL’, full precision for the object is shown.
+#' @param offsets Include the UTC offsets in the formatting
+#' @param usetz Include the time zone in the formatting
+#' @param precision The amount of precision: either "year", "month", "day", "hour", "minute", "second", or "nanosecond".
+#'        If `NULL` then full precision for the object is shown.
 #' @param ... Ignored
 #' @return A character vector
 #' @examples
@@ -25,6 +24,8 @@
 #'   format_iso8601(as_datetimeoffset("2020-05"))
 #'   format_iso8601(as_datetimeoffset("2020-05-10 20:15"))
 #'   format_iso8601(as_datetimeoffset("2020-05-10 20:15:05-07"))
+#'   if (require("lubridate"))
+#'     lubridate::format_ISO8601(as_datetimeoffset("2020-05-10 20:15:05-07"))
 #'
 #'   # pdfmark datetimes
 #'   format_pdfmark(as_datetimeoffset("2020-05"))
@@ -41,6 +42,14 @@
 #'     dt <- as_datetimeoffset(Sys.time())
 #'     format_nanotime(dt, format = "%F %H:%M:%E7S %Ez") # SQL Server datetimeoffset
 #'   }
+#'
+#'   # EDTF style formatting
+#'   format_edtf(as_datetimeoffset("2020-05"))
+#'   format_edtf(as_datetimeoffset("2020-05-10T20:15:05-07"))
+#'   dt <- datetimeoffset(2020, NA_integer_, 10)
+#'   format_edtf(dt)
+#'   # Not a valid EDTF string but useful serialization of `datetimeoffset()`
+#'   format_edtf(dt, precision = "nanosecond", usetz = TRUE)
 NULL
 
 #' @rdname format
@@ -60,8 +69,9 @@ format.datetimeoffset <- function(x, ...) {
            offset_str)
 }
 
-my_format_nanosecond <- function(ns) {
+my_format_nanosecond <- function(ns, edtf = FALSE, blank = FALSE) {
     s <- character(length(ns))
+    if (blank) return(s)
     idx <- which(!is.na(ns))
     if (length(idx > 0L)) {
         s_ns <- formatC(ns[idx], format = "d", flag = "0", width = 9L)
@@ -70,27 +80,21 @@ my_format_nanosecond <- function(ns) {
         s_ns <- ifelse(s_ns == "", "0", s_ns)
         s[idx] <- paste0(".", s_ns)
     }
+    idx <- which(is.na(ns))
+    if (edtf && length(idx))
+        s[idx] <- rep_len(".XXXXXXXXX", length(idx))
     s
 }
 
 #' @rdname format
+#' @param sep UTC offset separator.  Either ":" or "".
 #' @export
-format_iso8601 <- function(x, usetz = TRUE, precision = NULL, ...) {
-    precision <- precision %||% "ymdhmsn"
-    stopifnot(precision %in% c("y", "ym", "ymd", "ymdh", "ymdhm", "ymdhms", "ymdhmsn"))
-    if (precision == "y")
-        x <- set_month(x, NA_integer_)
-    if (precision == "ym")
-        x <- set_day(x, NA_integer_)
-    if (precision == "ymd")
-        x <- set_hour(x, NA_integer_)
-    if (precision == "ymdh")
-        x <- set_minute(x, NA_integer_)
-    if (precision == "ymdhm")
-        x <- set_second(x, NA_integer_)
-    if (precision == "ymdhms")
-        x <- set_nanosecond(x, NA_integer_)
-    if (isFALSE(usetz)) {
+format_iso8601 <- function(x, offsets = TRUE, precision = NULL, sep = ":", ...) {
+    precision <- precision %||% "nanosecond"
+    stopifnot(precision %in% c("year", "month", "day", "hour", "minute", "second", "nanosecond"))
+    stopifnot(sep %in% c(":", ""))
+    x <- datetime_narrow(x, precision)
+    if (isFALSE(offsets)) {
         x <- set_hour_offset(x, NA_integer_)
         x <- set_minute_offset(x, NA_integer_)
         x <- set_tz(x, NA_character_)
@@ -103,7 +107,7 @@ format_iso8601 <- function(x, usetz = TRUE, precision = NULL, ...) {
     minute_str <- my_format(field(x, "minute"), prefix = ":")
     second_str <- my_format(field(x, "second"), prefix = ":")
     nanosecond_str <- my_format_nanosecond(field(x, "nanosecond"))
-    offset_str <- my_format_tz(x)
+    offset_str <- my_format_tz(x, sep = sep)
     paste0(year_str, month_str, day_str,
            hour_str, minute_str, second_str, nanosecond_str,
            offset_str)
@@ -144,6 +148,35 @@ format_nanotime <- function(x, format = "%Y-%m-%dT%H:%M:%E9S%Ez", tz = get_tz(x)
     purrr::pmap_chr(df, base::format)
 }
 
+#' @rdname format
+#' @export
+format_edtf <- function(x, offsets = TRUE, precision = NULL, usetz = FALSE, ...) {
+    purrr::map_chr(x, format_edtf_helper, offsets = offsets, precision = precision, usetz = usetz)
+}
+
+format_edtf_helper <- function(x, offsets, precision, usetz) {
+    precision <- precision %||% datetime_precision(x, unspecified = TRUE)
+    stopifnot(precision %in% c("year", "month", "day", "hour", "minute", "second", "nanosecond"))
+    x <- datetime_narrow(x, precision)
+    if (isFALSE(offsets)) {
+        x <- set_hour_offset(x, NA_integer_)
+        x <- set_minute_offset(x, NA_integer_)
+        x <- set_tz(x, NA_character_)
+    }
+    precision <- dto_precision_integer(precision)
+    year_str <- my_format(field(x, "year"), width = 4L, edtf = TRUE, blank = precision < 1L)
+    month_str <- my_format(field(x, "month"), prefix = "-", edtf = TRUE, blank = precision < 2L)
+    day_str <- my_format(field(x, "day"), prefix = "-", edtf = TRUE, blank = precision < 3L)
+    hour_str <- my_format(field(x, "hour"), prefix = "T", edtf = TRUE, blank = precision < 4L)
+    minute_str <- my_format(field(x, "minute"), prefix = ":", edtf = TRUE, blank = precision < 5L)
+    second_str <- my_format(field(x, "second"), prefix = ":", edtf = TRUE, blank = precision < 6L)
+    nanosecond_str <- my_format_nanosecond(field(x, "nanosecond"), edtf = TRUE, blank = precision < 7L)
+    offset_str <- my_format_tz(x, edtf = offsets, add_tz = usetz)
+    paste0(year_str, month_str, day_str,
+           hour_str, minute_str, second_str, nanosecond_str,
+           offset_str)
+}
+
 update_nas <- function(x, pdfmark = FALSE) {
     # No smaller time units if missing bigger time units
     x <- set_day(x, ifelse(is.na(get_month(x)), NA_integer_, get_day(x)))
@@ -180,21 +213,28 @@ as_ymd_hms_str <- function(x, ...) {
     paste0(year_str, month_str, day_str, hour_str, minute_str, second_str)
 }
 
-my_format <- function(x, prefix = "", width = 2L, flag = "0") {
+my_format <- function(x, prefix = "", width = 2L, flag = "0", edtf = FALSE, blank = FALSE) {
     s <- character(length(x))
+    if (blank) return(s)
     idx <- which(!is.na(x))
     if (length(idx))
         s[idx] <- paste0(prefix, formatC(x[idx], format = "d", flag = flag, width = width))
+    idx <- which(is.na(x))
+    if (edtf && length(idx))
+        s[idx] <- paste0(prefix, strrep("X", width))
     s
 }
 
-my_format_tz <- function(x, sep = ":", no_zulu = FALSE, add_tz = FALSE) {
+my_format_tz <- function(x, sep = ":", no_zulu = FALSE, edtf = FALSE, add_tz = FALSE) {
     tz <- field(x, "tz")
     hour_offset <- field(x, "hour_offset")
     minute_offset <- field(x, "minute_offset")
+    hos <- my_format(hour_offset, width = 3L, flag = "0+")
+    hos <- ifelse(is.na(hour_offset) & ((!is.na(minute_offset) & edtf) | (edtf & add_tz)), "+XX", hos)
+    mos <- my_format(minute_offset, prefix = sep)
+    mos <- ifelse(is.na(minute_offset) & edtf & add_tz, ":XX", mos)
 
-    s <- paste0(my_format(hour_offset, width = 3L, flag = "0+"),
-                my_format(minute_offset, prefix = sep))
+    s <- paste0(hos, mos)
 
     id_zulu <- which(is_utc(tz))
     if (length(id_zulu) > 0L) {
@@ -223,5 +263,8 @@ my_format_tz <- function(x, sep = ":", no_zulu = FALSE, add_tz = FALSE) {
         }
         s[id_tz] <- s_offsets
     }
+    idx <- which(is.na(tz))
+    if (edtf && add_tz && length(idx))
+        s[idx] <- paste0(s[idx], "[X]")
     s
 }